backend/requirements.txt
asgiref==3.5.2
Django==3.2.12
gunicorn==20.1.0
pytz==2022.2.1
sqlparse==0.4.2
tzdata==2022.2
google-cloud-vision==3.1.2
djangorestframework==3.13.1
drf-yasg==1.21.4